+++ RELEASE_NOTES.md
@@ -4,10 +4,11 @@ This file contains one line summaries of commits that are worthy of mentioning i
 A new header is inserted each time a *tag* is created.
 
 ## main branch
-- backend: add support for GGP platform
 
+- backend: add support for GGP platform
 - gltfio: calculate primitive's AABB correctly. 
 - gltfio: fix glTF breaking issue
+- gltfio: recompute bounding boxes with morph targets
 - engine: primitives with `CullingMode::FRONT_AND_BACK` are now skipped.
 
 ## v1.28.2